добавлена функция Share Vacancy
