+++ app/src/main/java/ru/practicum/android/diploma/di/RepositoryModule.kt
@@ -6,8 +6,8 @@ import ru.practicum.android.diploma.favorites.data.impl.FavouriteVacancyReposito
 import ru.practicum.android.diploma.favorites.domain.db.FavouriteVacancyRepository
 import ru.practicum.android.diploma.search.data.repo.SearchRepositoryImpl
 import ru.practicum.android.diploma.search.domain.api.SearchRepository
-import ru.practicum.android.diploma.sharingandcontactvacancy.data.ExternalNavigatorImpl
-import ru.practicum.android.diploma.sharingandcontactvacancy.domain.api.ExternalNavigator
+import ru.practicum.android.diploma.vacancydetails.data.repo.ExternalNavigatorImpl
+import ru.practicum.android.diploma.vacancydetails.domain.api.ExternalNavigator
 import ru.practicum.android.diploma.vacancydetails.data.repo.DetailsRepositoryImpl
 import ru.practicum.android.diploma.vacancydetails.domain.api.DetailsRepository
