--- app/src/main/java/ru/practicum/android/diploma/sharingandcontactvacancy/domain/api/SharingContactInteractor.kt
@@ -1,9 +1,0 @@
-package ru.practicum.android.diploma.sharingandcontactvacancy.domain.api
-
-import android.content.Intent
-
-interface SharingContactInteractor {
-    fun shareVacancy(): Intent
-    fun contactVacancyByPhone(): Intent
-    fun contactVacancyByEmail(): Intent
-}
