+++ app/src/main/java/ru/practicum/android/diploma/vacancydetails/data/repo/ExternalNavigatorImpl.kt
@@ -1,0 +1,14 @@
+package ru.practicum.android.diploma.vacancydetails.data.repo
+
+import android.content.Intent
+import android.net.Uri
+import ru.practicum.android.diploma.vacancydetails.domain.api.ExternalNavigator
+
+class ExternalNavigatorImpl : ExternalNavigator {
+    override fun shareVacancy(link: String): Intent {
+        val shareIntent = Intent(Intent.ACTION_SEND)
+        shareIntent.setType("text/plain")
+        shareIntent.putExtra(Intent.EXTRA_TEXT, link)
+        return Intent.createChooser(shareIntent, "")
+    }
+}
