+++ app/src/main/java/ru/practicum/android/diploma/vacancydetails/domain/api/ExternalNavigator.kt
@@ -1,0 +1,7 @@
+package ru.practicum.android.diploma.vacancydetails.domain.api
+
+import android.content.Intent
+
+interface ExternalNavigator {
+    fun shareVacancy(link: String): Intent
+}
