+++ app/src/main/res/layout/fragment_vacancy_details.xml
@@ -55,6 +55,7 @@
             android:layout_marginTop="20dp"
             android:layout_marginEnd="28dp"
             android:clickable="true"
+            android:foreground="?attr/selectableItemBackground"
             android:src="@drawable/sharing_24px_button"
             app:layout_constraintEnd_toStartOf="@+id/favoriteVacansyIv"
             app:layout_constraintTop_toTopOf="parent" />
